+++ src/SubStatement.php
@@ -97,19 +97,6 @@ final class SubStatement extends StatementObject
         return $subStatement;
     }
 
-    /**
-     * @deprecated since 1.2, to be removed in 3.0
-     */
-    public function withTimestamp(\DateTime $timestamp = null)
-    {
-        @trigger_error(sprintf('The "%s()" method is deprecated since 1.2 and will be removed in 3.0. Use "%s::withCreated()" instead.', __METHOD__, __CLASS__), E_USER_DEPRECATED);
-
-        $statement = clone $this;
-        $statement->created = $timestamp;
-
-        return $statement;
-    }
-
     public function withCreated(\DateTime $created = null)
     {
         $statement = clone $this;
@@ -177,21 +164,6 @@ final class SubStatement extends StatementObject
     public function getResult()
     {
         return $this->result;
-    }
-
-    /**
-     * Returns the timestamp of when the events described in this statement
-     * occurred.
-     *
-     * @return \DateTime The timestamp
-     *
-     * @deprecated since 1.2, to be removed in 3.0
-     */
-    public function getTimestamp()
-    {
-        @trigger_error(sprintf('The "%s()" method is deprecated since 1.2 and will be removed in 3.0. Use "%s::getCreated()" instead.', __METHOD__, __CLASS__), E_USER_DEPRECATED);
-
-        return $this->created;
     }
 
     /**